+++ src/StructLinq/Take/RefTakeEnumerator.cs
@@ -8,10 +8,10 @@ namespace StructLinq.Take
         private TEnumerator enumerator;
         private readonly int endIndex;
         private int index;
-        public RefTakeEnumerator(ref TEnumerator enumerator, uint length)
+        public RefTakeEnumerator(ref TEnumerator enumerator, int length)
         {
             this.enumerator = enumerator;
-            endIndex = (int)length - 1;
+            endIndex = length - 1;
             index = -1;
         }
